+++ dwtools/Ltas_extensions.cpp
@@ -24,20 +24,20 @@ void Ltas_fitTiltLine (Ltas me, double fmin, double fmax, bool lnf, int method, 
 		if (fmax <= fmin) {
 			fmin = my xmin; fmax = my xmax;
 		}
-		long ifmin, ifmax, numberOfSamples = Sampled_getWindowSamples (me, fmin, fmax, &ifmin, &ifmax);
+		integer ifmin, ifmax, numberOfSamples = Sampled_getWindowSamples (me, fmin, fmax, & ifmin, & ifmax);
 		if (numberOfSamples < 2) {
 			Melder_throw (U"There must be at least two data points to fit a line.");
 		}
 		autoNUMvector<double> x (1, numberOfSamples);
 		autoNUMvector<double> y (1, numberOfSamples);
-		for (long i = ifmin; i <= ifmax; i++) {
-			long ixy = i - ifmin + 1;
-			x[ixy] = my x1 + (i - 1) * my dx;
+		for (integer i = ifmin; i <= ifmax; i ++) {
+			integer ixy = i - ifmin + 1;
+			x [ixy] = my x1 + (i - 1) * my dx;
 			if (lnf) {
 				// For Ltas always x1 > 0
-				x[ixy] = log10 (x[ixy]);
+				x [ixy] = log10 (x [ixy]);
 			}
-			y[ixy] = my z[1][i];
+			y [ixy] = my z [1] [i];
 		}
 		NUMlineFit (x.peek(), y.peek(), numberOfSamples, a, b, method);
 	} catch (MelderError) {
